Cleanup codes
- Remove IDOrSelector
- Move NameOrSelector into resource_selector.go
- Rename definition_types.go to string_or_filepath.go
