+++ core/resource_selector.go
@@ -17,6 +17,7 @@ package core
 import (
 	"fmt"
 
+	"github.com/goccy/go-yaml"
 	"github.com/sacloud/autoscaler/validate"
 	"github.com/sacloud/libsacloud/v2/sacloud"
 	"github.com/sacloud/libsacloud/v2/sacloud/search"
@@ -96,5 +97,22 @@ func (rs *ResourceSelector) Validate() error {
 	if !rs.ID.IsEmpty() && (len(rs.Names) > 0 || len(rs.Tags) > 0) {
 		return fmt.Errorf("selector.ID and (selector.Names or selector.Tags): cannot specify both")
 	}
+	return nil
+}
+
+// NameOrSelector 名前(文字列)、もしくはResourceSelectorを表すstruct
+type NameOrSelector struct {
+	ResourceSelector
+}
+
+func (v *NameOrSelector) UnmarshalYAML(data []byte) error {
+	// セレクタとしてUnmarshalしてみてエラーだったら文字列と見なす
+	var selector ResourceSelector
+	if err := yaml.UnmarshalWithOptions(data, &selector, yaml.Strict()); err != nil {
+		selector = ResourceSelector{
+			Names: []string{string(data)},
+		}
+	}
+	*v = NameOrSelector{ResourceSelector: selector}
 	return nil
 }
